requirements.txt
python-telegram-bot==11.1.0
firebase-admin==4.4.0
spotipy==2.16.0
pymaybe==0.1.6